bt-unit13

+++ app/Http/Controllers/Backend/UserController.php
@@ -12,6 +12,7 @@ use Illuminate\Support\Facades\Storage;
 class UserController extends Controller
 {
     public function index(){
+        $this->authorize('isQtvOrAdmin', User::class);
     	$users = User::withTrashed()->orderBy('role','DESC')->paginate(5);
     	return view('backend.users.index')->with([
     		'users' => $users
@@ -19,10 +20,12 @@ class UserController extends Controller
     }
 
     public function create(){
+        $this->authorize('isAdmin', User::class);
         return view('backend.users.create');
     }
 
     public function show(Request $request,$id){
+        $this->authorize('isQtvOrAdmin', User::class);
         $user = User::withTrashed()->where('id',$id)->first();
         return view('backend.users.detail')->with([
             'user' => $user
@@ -30,9 +33,20 @@ class UserController extends Controller
     }
 
      public function store(RequestUser $requestUser){
+        $this->authorize('isAdmin', User::class);
         $user = new User();
         $user->name = $requestUser->get('name');
         $user->email = $requestUser->get('email');
+
+        if($requestUser->hasFile('avatar')){
+            $image = $requestUser->file('avatar');
+            $name_image = date('YmdHis')."_".$image->getClientOriginalName();
+            Storage::disk('public')->putFileAs('images/user_avatar/',$image,$name_image);
+            $user->avatar = 'storage/images/user_avatar/'.$name_image;
+        }else{
+            $user->avatar = 'storage/images/user_avatar/default-avatar.jpg';
+        }
+
         $user->password = bcrypt($requestUser->get('password'));
         $user->address = $requestUser->get('address');
         $user->phone = $requestUser->get('phone');
@@ -41,24 +55,27 @@ class UserController extends Controller
         return redirect()->route('backend.user.index');
     }
 
-    public function destroy($id){
-        $user = User::withTrashed()->where('id',$id)->first();
+    public function forceDelete($id){
+        $this->authorize('isAdmin', User::class);
+        $user = User::onlyTrashed()->where('id',$id)->get();
         $user->forceDelete();
         return redirect()->back();
     }
 
     public function editStatus($id){
+        $this->authorize('isAdmin', User::class);
     	$user = User::withTrashed()->where('id',$id)->first();
     	if ($user->status == 1) {
     		$user->status = 0;
     	}else{
     		$user->status = 1;
     	}
-    	$user->save();
+    	$user->update();
     	return redirect()->route('backend.user.index');
     }
 
     public function openOrBlock($id){
+        $this->authorize('isAdmin', User::class);
     	$user = User::withTrashed()->where('id',$id)->first();
     	if ($user->trashed()) {
     		$user->restore();
@@ -70,16 +87,15 @@ class UserController extends Controller
     }
 
     public function showProducts(Request $request,$id){
+        $this->authorize('isQtvOrAdmin', User::class);
         $products = User::withTrashed()->where('id',$id)->first()->products()->orderBy('updated_at','DESC')->paginate(4);
         $categories=Category::all();
-        //dd($request->all());
         if($request->name) $products->where('name','like','%'.$request->name.'%');
         if($request->category_id) $products->where('category_id',$request->category_id);
-        //$products = $products->orderBy('updated_at','DESC')->paginate(4);
-        // dd($products);
         return view('backend.products.index')->with([
             'products' => $products,
             'categories' => $categories
         ]);
     }
+
 }
